Admin Controller CleanUp

+++ app/Http/Controllers/Api/AdminController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\User;
+use Illuminate\Support\Facades\Auth;
 
 class AdminController extends Controller
 {
@@ -16,14 +17,14 @@ class AdminController extends Controller
     public function index()
     {
         //Check if user is Admin
-        
+
         if (Auth::check() && Auth::user()->role == 2) {
 
             // Fetch completed requests
 
             $completed_requests = DB::table('users')->where('isFunded', '=', 1)->get();
 
-        // Count completed requests
+            // Count completed requests
 
             $count_completed = count($completed_requests);
             return response()->json([
@@ -96,24 +97,30 @@ class AdminController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function destroy($id)
     {
-        
+
         //delete User account
-        if(User::where('id', $id)->exists()) {
-            $user = User::find($id);
-            $user->delete();
-    
+        if (Auth::check() && Auth::user()->role == 2) {
+            if(User::where('id', $id)->exists()) {
+                $user = User::find($id);
+                $user->delete();
+
+                return response()->json([
+                    "message" => "User account successfully deleted"
+                ], 202);
+            } else {
+                return response()->json([
+                    "message" => "User account not found"
+                ], 404);
+            }
+        } else {
             return response()->json([
-              "message" => "User account successfully deleted"
-            ], 202);
-          } else {
-            return response()->json([
-              "message" => "User account not found"
-            ], 404);
-          }
+                'message' => 'You do not have permission to perform this action'
+            ]);
+        }
     }
-    
+
 }
